% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{add_id_col}
\alias{add_id_col}
\alias{listindex_to_col}
\title{Add id column to dataframe}
\usage{
add_id_col(.data, value = "1", var = "doc_id")

listindex_to_col(.data, var = "doc_id")
}
\arguments{
\item{.data}{A list (or list of lists) of data frames.}

\item{value}{A value to assign as the id.}

\item{var}{Character name of the new id column.}
}
\description{
\code{add_id_col()} adds an id column to either a data frame or to each data frame in a list.

\code{listindex_to_col()} adds the listindex from the first list of a list of data frames
as an ID column in each of the nested data frames.
}
\seealso{
\code{\link{unnest_df}}
}
