% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spec.R
\name{add_missing_spec_tables}
\alias{add_missing_spec_tables}
\title{Add missing tables from the specification}
\usage{
add_missing_spec_tables(table_list, table_spec, checked_spec)
}
\arguments{
\item{table_list}{A list of data frames to check against the spec.}

\item{table_spec}{The table spec. See details.}

\item{checked_spec}{A return from \code{\link{check_spec}()}.}
}
\description{
Add missing tables from the specification
}
\seealso{
Other Data Spec Functions: 
\code{\link{add_missing_spec_cols}()},
\code{\link{assert_case}()},
\code{\link{change_case_from_spec}()},
\code{\link{check_spec}()},
\code{\link{coerce_to_spec}()},
\code{\link{data_case}()}
}
\concept{Data Spec Functions}
