% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{as_int}
\alias{as_int}
\title{Coerces to integer without warning}
\usage{
as_int(x, na_regex = "[^0-9]")
}
\arguments{
\item{x}{object to coerce silently to integer.}

\item{na_regex}{regular expression of values to convert to \code{NA}.}
}
\description{
\lifecycle{questioning}
}
