% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spec.R
\name{assert_case}
\alias{assert_case}
\title{Assert the data case onto the object}
\usage{
assert_case(x, target_case = "snake", .table_spec = data_spec(x))
}
\arguments{
\item{x}{An object to assert case.}

\item{target_case}{The case to transform it to.}

\item{.table_spec}{A file specification list. Defaults to deriving from the object.}
}
\description{
\code{assert_case()} checks an object for what case the tables and fields are in
and converts them to the \code{target_case} us required.\cr
\cr
Requires the object to have the \code{data_case} attribute.
}
\seealso{
Other Data Spec Functions: 
\code{\link{add_missing_spec_cols}()},
\code{\link{add_missing_spec_tables}()},
\code{\link{change_case_from_spec}()},
\code{\link{check_spec}()},
\code{\link{coerce_to_spec}()},
\code{\link{data_case}()}
}
\concept{Data Spec Functions}
