% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg_suggests.R
\name{check_pkg_suggests}
\alias{check_pkg_suggests}
\title{Check if Packages are Installed}
\usage{
check_pkg_suggests(pkg)
}
\arguments{
\item{pkg}{A vector of character package names.}
}
\description{
\code{check_pkg_suggests()} checks to see if packages are installed. This can be
used to check for the existance of packages in the Suggests line.
}
\examples{
\dontrun{
check_pkg_suggests(c("dplyr", "tidyr"))
check_pkg_suggests(c("dplyr", "tidyr", "notrealpackage", "alsofake"))
}
}
