% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_by.R
\name{clean_by}
\alias{clean_by}
\title{Clean "by" argument}
\usage{
clean_by(by)
}
\arguments{
\item{by}{A character vector of variables to join by. Same specification as used in
\code{\link[dplyr]{join}}.\cr
\cr
To join by different variables on \code{df} and \code{df_sql} use a named vector.
For example, by = c("a" = "b") will match \code{df.a} to \code{df_sql.b}.}
}
\description{
\code{clean_by()} takes a valid \code{by} input of a character vector (named or unnamed) and transforms
it into a list. Similar to \code{\link[dplyr]{common_by}} but does not require the actual data frames.
}
