% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{col_rep}
\alias{col_rep}
\title{Simple column specifications}
\usage{
col_rep(str, spec_type = "readr")
}
\arguments{
\item{str}{String specifying the expansion.}

\item{spec_type}{String. Either "readr" for short letter notation (e.g.,
\code{ccnnndcc}) or "readxl" for long vector notation
(e.g., \code{c("text", "text", "numeric")})}
}
\value{
An expanded string or a vector of column type specifications.
}
\description{
\code{col_rep()} allows for a simple specification of column types for use in read
functions. Enter in a letter followed by a number to repeat that type that number
of times. \cr
\cr
Supports output for use in \code{readr} and \code{readxl}, and other functions
that take column type specifications in the same format.
}
\examples{
col_rep("c3n2dl3-2c")

col_rep("c3n2dl3-2c", "readxl")

}
