% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spec.R
\name{data_case}
\alias{data_case}
\alias{data_spec}
\title{Attribute access functions}
\usage{
data_case(x)

data_spec(x)
}
\arguments{
\item{x}{An object to check.}
}
\description{
Returns the namesake attribute.
}
\seealso{
Other Data Spec Functions: 
\code{\link{add_missing_spec_cols}()},
\code{\link{add_missing_spec_tables}()},
\code{\link{assert_case}()},
\code{\link{change_case_from_spec}()},
\code{\link{check_spec}()},
\code{\link{coerce_to_spec}()}
}
\concept{Data Spec Functions}
