% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_case_from_spec.R
\name{data_model_to_snake}
\alias{data_model_to_snake}
\title{Convert table names}
\usage{
data_model_to_snake(table_list, table_spec)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{data_model_to_snake()} was renamed to \code{native_to_snake_case()} to make the function
general in name.
}
\keyword{internal}
