% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_mod.R
\name{distinct_insensitive}
\alias{distinct_insensitive}
\title{Insensitive dplyr distinct}
\usage{
distinct_insensitive(.data, ..., .keep_all = FALSE)
}
\arguments{
\item{.data}{See \code{\link[dplyr]{distinct}}.}

\item{...}{See \code{\link[dplyr]{distinct}}.}

\item{.keep_all}{See \code{\link[dplyr]{distinct}}.}
}
\description{
Alternative to \code{\link[dplyr]{distinct}} to operate insensitive to case. Parameters are defined identically to the
dplyr base function.
}
\seealso{
\code{\link[dplyr]{distinct}} from \code{dplyr} for parameter definitions and function.
}
