% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json_zip.R
\name{fix_json_null}
\alias{fix_json_null}
\title{Fix 'null' values in JSON}
\usage{
fix_json_null(json_object)
}
\arguments{
\item{json_object}{An object read in from \code{read_json()}.}
}
\value{
A corrected tibble with \code{NA} in place of the null.
}
\description{
\code{fix_json_null()} replaces 'null' values in the JSON with an NA. This
is needed for single row tables in the JSON with a null value. In this case,
\code{\link[jsonlite]{read_json}(..., simplifyVector = FALSE)} returns a list instead
of a data.frame because of the 'null'. This function takes that list, replaces the 'null'
with an \code{NA} and returns the data.frame.
}
