% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flatten_data}
\alias{flatten_data}
\title{Create a flat file from multiple data frames}
\usage{
flatten_data(x, ...)
}
\arguments{
\item{x}{A list of one or more collections of data frames to be flattened.}

\item{...}{Arguments passed on to methods.}
}
\value{
A data frame with the flat file representation of the input data.
}
\description{
A generic function that is used to create a single flat file from a list of data
frames. This list is usually created by reading in a data format
with data spanning multiple tables.
}
