% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_contract_num.R
\name{format_contract}
\alias{format_contract}
\title{Format contract number}
\usage{
format_contract(contract, hyphenated = TRUE, master_only = FALSE)
}
\arguments{
\item{contract}{Character string, contract number.}

\item{hyphenated, master_only}{Logical, options for controlling the \emph{output}.
\code{master_only} returns only the parent contract.}
}
\value{
A character string in the specified format.
}
\description{
\code{format_contract()} takes in a single contract number and returns a consistently-formatted
contract number, according to user controls for the output.
}
\examples{
# control the output with the two logical parameters
format_contract("FA8616-14-D-6060:0001",
                hyphenated = TRUE,
                master_only = FALSE)
format_contract("FA8620-15-G-4040-0039",
                hyphenated = TRUE,
                master_only = TRUE)
format_contract("FA8620-15-G-4040-FA8620-15-F-4040",
                hyphenated = FALSE,
                master_only = TRUE)
format_contract("FA862015C4040",
                hyphenated = FALSE,
                master_only = FALSE)

}
