% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions_excel.R
\name{get_excel_tables}
\alias{get_excel_tables}
\title{List Excel tables in a workbook}
\usage{
get_excel_tables(wb, sheets = NULL)
}
\arguments{
\item{wb}{A workbook object}

\item{sheets}{Character vector of sheet names to list tables from. If \code{NULL},
lists all sheets.}
}
\value{
A tibble with the worksheet, table name, and cell references.
}
\description{
\code{get_excel_tables()} is similar to \code{\link[openxlsx]{getTables}()} but
returns more information.
}
\examples{
example_file <- system.file("examples/excel_examples.xlsx", package = "costmisc")

wb <- openxlsx::loadWorkbook(example_file)
get_excel_tables(wb)

}
\seealso{
Other Read Excel tables: 
\code{\link{read_excel_table}()}
}
\concept{Read Excel tables}
