% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_mod.R
\name{insensitive}
\alias{insensitive}
\title{Insensitive dplyr joins}
\usage{
insensitive(fun = dplyr::inner_join)
}
\arguments{
\item{fun}{Function from \code{dplyr::join} family.}
}
\description{
Function to wrap around the dplyr joins to make them insensitive to case. Author credit Jim Hester.\cr
\cr
\url{https://gist.github.com/jimhester/a060323a05b40c6ada34}
}
\examples{
\donttest{insensitive(left_join)(x, y)}

}
\seealso{
\code{\link[dplyr]{join}} and related functions from \code{dplyr}.
}
