% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{is_scalar_numeric}
\alias{is_scalar_numeric}
\title{Check is scalar type number}
\usage{
is_scalar_numeric(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\lifecycle{questioning}
This function was moved from \code{costmodelr}. The use case is unclear and as such this
function may be removed from a later version.\cr
\cr
\code{is_scalar_numeric()} checks for a given type and whether the vector is
"scalar", that is, of length 1.
}
\seealso{
\code{\link[rlang]{scalar-type-predicates}}
}
