% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{listindex_to_col}
\alias{listindex_to_col}
\title{Add id column to a list}
\usage{
listindex_to_col(.nestedlist)
}
\arguments{
\item{.nestedlist}{A nested list of data frames.}
}
\description{
\code{listindex_to_col()} adds the listindex from the first list of a list of data frames
as an ID column in each of the nested data frames.
}
\examples{
files <- system.file("extdata/multiple-flexfiles", package = "csdrtools")

flexfiles <- read_folder(files, read_ff) \%>\%
listindex_to_col()

}
