% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_difftime.R
\name{paste_difftime}
\alias{paste_difftime}
\title{Paste function for time difference}
\usage{
paste_difftime(
  x,
  time_msg = "Time difference of",
  digits = getOption("digits")
)
}
\arguments{
\item{x}{A timediff object.}

\item{time_msg}{The message prefix.}

\item{digits}{Number of digits option.}
}
\description{
\code{paste_difftime()} creates a message for an elapsed time. Taken from \code{print.difftime()}.
}
