% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_folder}
\alias{read_folder}
\title{Read of a folder of files}
\usage{
read_folder(
  folder,
  read_function,
  .clean_file_names = TRUE,
  .id = NULL,
  .recursive = TRUE,
  ...
)
}
\arguments{
\item{folder}{A folder path to read.}

\item{read_function}{The function to use to read each file.}

\item{.clean_file_names}{Logical to clean names into snake_case or not.}

\item{.id}{Optionally add an id variable to each file table as this name}

\item{.recursive}{Logical to recursively load the folder.}

\item{...}{Arguments passed to \code{read_function}.}
}
\value{
A list of output from the \code{read_function}.
}
\description{
\code{read_folder()} is a wrapper around \code{lapply} to read an entire folder of files
into a list. No file checking is implemented, so each file in the target folder must be
of the same type.
}
