% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_case_from_spec.R
\name{snake_to_data_model}
\alias{snake_to_data_model}
\title{Convert table names}
\usage{
snake_to_data_model(table_list, table_spec)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{snake_to_data_model()} was renamed to \code{snake_to_native_case()} to make the function
general in name.
}
\keyword{internal}
