% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{strip_attributes}
\alias{strip_attributes}
\title{Strip attributes from object}
\usage{
strip_attributes(x, .retain_class = TRUE)
}
\arguments{
\item{x}{object to strip attributes from.}

\item{.retain_class}{logical whether or not to retain the class.}
}
\description{
\lifecycle{questioning}
}
