% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{unnest_df}
\alias{unnest_df}
\title{Internal function to support stacking FlexFiles}
\usage{
unnest_df(.data)
}
\arguments{
\item{.data}{A list of FlexFile submissions' tibbles converted from JSON format.}
}
\value{
A list of stacked tibbles of multiple dataframes
}
\description{
\code{unnest_df()} supports FlexFile stacking Internal function.
}
