% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate a data set}
\usage{
validate_data(x, ...)
}
\arguments{
\item{x}{A data object to validate.}

\item{...}{Arguments passed on to methods.}
}
\description{
A generic function that is used to validate data from a specialized read function
within the costverse.
}
