% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json_zip.R
\name{write_json_zip}
\alias{write_json_zip}
\title{Write a list of data.frames to zipped JSON}
\usage{
write_json_zip(x, path, ...)
}
\arguments{
\item{x}{A list of named data.frames to write to a zipped folder.}

\item{path}{Path for the new .zip file (should include ".zip" at the end).}

\item{...}{Additional arguments passed to \code{\link[jsonlite]{write_json}()}.}
}
\description{
\code{write_json_zip()} will write each data.frame in a named list to individual JSON files
inside of a zipped folder.
}
\seealso{
Other JSON list functions: 
\code{\link{read_json_zip}()}
}
\concept{JSON list functions}
