% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{add_id_col}
\alias{add_id_col}
\title{Add id column to dataframe}
\usage{
add_id_col(.data, value = "1", var = "ff_id")
}
\arguments{
\item{.data}{A list of FlexFile submissions' tibbles converted from JSON format.}

\item{value}{A value to assign as the id.}

\item{var}{Character name of which to name the id column.}
}
\description{
\code{add_id_col()} adds an id column to either a data frame or to each data frame in a list.
}
\examples{
file <- system.file("extdata", "Sample File_FF.zip", package = "csdrtools")

flexfile <- read_ff(file) \%>\%
add_id_col()

}
