% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spec.R
\name{check_spec}
\alias{check_spec}
\title{Check file specification}
\usage{
check_spec(table_list, table_spec, type_label = "Import File", .silent = TRUE)
}
\arguments{
\item{table_list}{A list of data frames to check against the spec.}

\item{table_spec}{The table spec. See details.}

\item{type_label}{A character label for the check message.}

\item{.silent}{Logical whether or not to print results to the console.}
}
\value{
A list containing the results of the check.
}
\description{
\code{check_spec()} will check a list of data against an expected file specification.
}
\seealso{
Other Data Spec Functions: 
\code{\link{add_missing_spec_cols}()},
\code{\link{add_missing_spec_tables}()},
\code{\link{change_case_from_spec}()},
\code{\link{coerce_to_spec}()}
}
\concept{Data Spec Functions}
