% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{unnest_df}
\alias{unnest_df}
\title{Unnest identical data frames}
\usage{
unnest_df(.data)
}
\arguments{
\item{.data}{A list of lists of dataframes.}
}
\value{
A list of stacked tibbles.
}
\description{
\code{unnest_df()} stacks a list of lists of dataframes. This assumes that each
level 2 list has the same dataframes with the same columns. This is used when
reading in a set of files with identical format.
}
\seealso{
\code{\link{listindex_to_col}}
}
